Rename: Plate관련 컴포넌트를 public/molecules 폴더로 이전
