+++ src/app/my_page/page.tsx
@@ -5,8 +5,8 @@ import MainSection from "@/components/public/molecules/MainSection/MainSection";
 import { ReactNode, RefObject, useRef, useState } from "react";
 import SettingElem from "@/components/my_page/molecules/SettingElem/SettingElem";
 import SettingSection from "@/components/my_page/organisms/SettingSection/SettingSection";
-import PlateFront from "@/components/public/atoms/PlateFront/PlateFront";
-import PlateBack from "@/components/public/atoms/PlateBack/PlateBack";
+import PlateFront from "@/components/public/molecules/PlateFront/PlateFront";
+import PlateBack from "@/components/public/molecules/PlateBack/PlateBack";
 import { TChinghoRank } from "@/types";
 import { IChingho } from "@/interfaces";
 import PlateChinghoBtn from "@/components/my_page/atoms/PlateChinghoBtn/PlateChinghoBtn";
